fix: Dockerfile install script relative path issue.

+++ create-symlinks.sh
@@ -1,13 +1,13 @@
 #!/bin/bash
 
-DOTFILES_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+DOTFILES_DIR="$HOME/.dotfiles"
 
 # Ensure directories exist.
 mkdir -p "$HOME/.config"
 mkdir -p "$HOME/.config/gh-copilot"
 mkdir -p "$HOME/.backup_dotfiles"
 
-# List of files to check
+# List of files to check.
 files=(
 	"$HOME/.config/nvim"
 	"$HOME/.config/gh-copilot/config.yml"
@@ -21,7 +21,7 @@ files=(
 # Backup any existing files that are not symlinks.
 echo "Backing up existing files..."
 for file in "${files[@]}"; do
-	if [ ! -L "$file" ]; then
+	if [ -e "$file" ] && [ ! -L "$file" ]; then
 		echo "Backing up '$file'"
 		mv "$file" "$HOME/.backup_dotfiles/"
 	fi
@@ -30,8 +30,8 @@ echo "...done! Backups were saved to $HOME/.backup_dotfiles/"
 
 # Create symlinks.
 echo "Creating symlinks to .dotfiles config files..."
-ln -sf "$DOTFILES_DIR/nvim" "$HOME/.config/nvim/"
-ln -sf "$DOTFILES_DIR/wezterm" "$HOME/.config/wezterm/"
+ln -sf "$DOTFILES_DIR/nvim" "$HOME/.config/nvim"
+ln -sf "$DOTFILES_DIR/wezterm" "$HOME/.config/wezterm"
 ln -sf "$DOTFILES_DIR/gh-copilot/config.yml" "$HOME/.config/gh-copilot/config.yml"
 ln -sf "$DOTFILES_DIR/.gitconfig" "$HOME/.gitconfig"
 ln -sf "$DOTFILES_DIR/.p10k.zsh" "$HOME/.p10k.zsh"
